fix: Fix rng generation and street mask projections
In a previous commit, self.rng was moved to be defined before the seed was determined, breaking the package. This moves it back to where it belongs. Additionally, Street's execute() function has to handle CRS' carefully due to OSM using 4326. A previous commit introduced a bug by using self.crs instead of 4326, and this fixes that. Finally, whitespace was cleaned up throughout the codebase.