Add street_k mask
This is a wrapper around the street mask that incrementally increases
the masking parameters until a given privacy target is met. Although
this comes at the cost of requiring population data.

When population data is available, it may make sense to use location
swapping instead, unless you want some of other benefits of street
masking (such as the reduced chance of false attribution, or reduced
chance of points being displaced across geographic barriers such as
rivers, mountains, or cliffs).